Update README document
Change the outline of the README document to represent the purpose of the
project in detail. The rewritten outline describe the aim of project as a
clone-coding project of Discord. It also explains that the project was
designed to learn the basics of using WPF and WCF, and how to use Git and
Github in a way that is actually used at work.